feat(lexiannot): 适配 MoviePilot 新版 LLM 助手，提取字幕助手类 SubtitleHelper (v1.2.6)
