+++ package.v2.json
@@ -588,11 +588,12 @@
     "name": "美剧生词标注",
     "description": "根据CEFR等级，为英语影视剧标注高级词汇。",
     "labels": "英语",
-    "version": "1.2.5",
+    "version": "1.2.6",
     "icon": "LexiAnnot.png",
     "author": "wumode",
     "level": 1,
     "history": {
+      "v1.2.6": "适配 MoviePilot 新版 LLM 助手",
       "v1.2.5": "langchain 1.x 兼容 (主程序版本需高于 2.9.17)",
       "v1.2.4": "增强数据校验",
       "v1.2.3": "优化提示词",
